+++ tomcat/src/main/java/org/apache/coyote/http11/message/HttpRequest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.message;
 
 import jakarta.servlet.http.HttpSession;
 import org.apache.catalina.RequestCookie;
