+++ tomcat/src/main/java/org/apache/coyote/http11/message/HttpResponse.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.message;
 
 public class HttpResponse {
