+++ tomcat/src/main/java/org/apache/coyote/http11/message/HttpResponseHeader.java
@@ -1,47 +1,44 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.message;
 
-import java.util.LinkedHashMap;
+import java.util.HashMap;
 import java.util.Map;
-import org.apache.catalina.RequestCookie;
+import org.apache.catalina.ResponseCookie;
 
-public class HttpRequestHeader {
+public class HttpResponseHeader {
 
     private final Map<String, String> values;
-    private RequestCookie cookie;
+    private ResponseCookie cookie;
 
-    public HttpRequestHeader(Map<String, String> values) {
-        this.values = values;
-    }
-
-    public HttpRequestHeader() {
-        this.values = new LinkedHashMap<>();
+    public HttpResponseHeader() {
+        this.values = new HashMap<>();
     }
 
     public void add(String key, String value) {
         this.values.put(key, value);
     }
 
-    public void addCookie(RequestCookie requestCookie) {
-        this.cookie = requestCookie;
+    public void addCookie(ResponseCookie cookie) {
+        this.cookie = cookie;
         this.values.remove("Cookie");
-    }
-
-    public boolean contains(String key) {
-        return values.containsKey(key);
     }
 
     public String get(String comparedKey) {
         return values.keySet().stream()
                 .filter(key -> key.equals(comparedKey))
                 .findAny()
+                .map(values::get)
                 .orElseThrow(() -> new IllegalArgumentException(comparedKey + "와 일치하는 값이 헤더에 존재하지 않습니다."));
     }
 
-    public boolean hasCookie() {
-        return this.values != null;
+    public Map<String, String> getValues() {
+        return new HashMap<>(values);
     }
 
-    public RequestCookie getCookie() {
+    public boolean hasCookie() {
+        return this.cookie != null;
+    }
+
+    public ResponseCookie getCookie() {
         return this.cookie;
     }
 }
