+++ tomcat/src/main/java/org/apache/coyote/http11/message/RequestLine.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.message;
 
 public record RequestLine(
         HttpMethod httpMethod,
