+++ tomcat/src/main/java/org/apache/coyote/http11/message/StatusLine.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.message;
 
 public record StatusLine(
         HttpStatus statusCode,
